Change finalization params to default to regtest (#438)

+++ src/blockchain/blockchain_parameters.cpp
@@ -31,7 +31,7 @@ Parameters BuildMainNetParameters() {
   p.maximum_supply = 2718275100 * UNIT;  // e billion UTE
   assert(p.maximum_supply == p.initial_supply + std::accumulate(p.reward_schedule.begin(), p.reward_schedule.end(), CAmount()) * p.period_blocks);
   p.reward_function = [](const Parameters &p, Height h) -> CAmount {
-    const int period = h / p.period_blocks;
+    const uint64_t period = h / p.period_blocks;
     if (period >= p.reward_schedule.size()) {
       return 0;
     }
